Created navigation.js file
